+++ output.c
@@ -6,19 +6,19 @@
 #include "hydro.h"
 
 
-/* double get_gamma_max(hydro_fields f, hydro_params p)
+/* float get_gamma_max(hydro_fields f, hydro_params p)
  *
  * Returns the largest zone-centred gamma factor found anywhere
  * in the simulation box.
  */
-double get_gamma_max(hydro_fields f, hydro_params p) {
+float get_gamma_max(hydro_fields f, hydro_params p) {
 #ifndef SCALAR
 
   int x, y, z, xmax;
 
-  double gmax = f.W[0][0][0];
+  float gmax = f.W[0][0][0];
   
-  double gtest;
+  float gtest;
 
   // Just search for maxmimum
   for(x = 1; x <= p.slicex; x++) {
@@ -41,18 +41,18 @@ double get_gamma_max(hydro_fields f, hydro_params p) {
 }
 
 
-/* double get_veltot(hydro_fields f, hydro_params p)
+/* float get_veltot(hydro_fields f, hydro_params p)
  *
  * The sum of the fluid (3-)velocity everywhere. A strange quantity
  * on its own, but allows calculation of average fluid velocity.
  */
-double get_veltot(hydro_fields f, hydro_params p) {
+float get_veltot(hydro_fields f, hydro_params p) {
 
 #ifndef SCALAR
   int x, y, z, xmax;
 
   
-  double veltot = 0.0;
+  float veltot = 0.0;
 
   // Just search for maxmimum
   for(x = 1; x <= p.slicex; x++) {
@@ -73,12 +73,12 @@ double get_veltot(hydro_fields f, hydro_params p) {
 }
 
 
-/* void dump(double *field, hydro_params p) 
+/* void dump(float *field, hydro_params p) 
  *
  * Dumps a field to stderr. Expects field to have N entries.
  * For debugging purposes...
  */
-void dump(double *field, hydro_params p) {
+void dump(float *field, hydro_params p) {
   int x;
 
   fprintf(stderr,"%g", field[0]);
@@ -90,20 +90,20 @@ void dump(double *field, hydro_params p) {
 
 
 
-/* void histo_field(double ***field, hydro_params p, int step)
+/* void histo_field(float ***field, hydro_params p, int step)
  *
  * Calculate a histogram of the field, and store in a file
  * labelled by the timestep.
  */
-void histo_field(double ***field, hydro_params p, int step) {
+void histo_field(float ***field, hydro_params p, int step) {
   int x, y, z;
 
-  double fmax = 0.0;
-  double fmin = 0.0;
+  float fmax = 0.0;
+  float fmin = 0.0;
 
-  double ftest;
+  float ftest;
 
-  double start = clock();
+  float start = clock();
 
 
   fmax = field[0][0][0];
@@ -124,7 +124,7 @@ void histo_field(double ***field, hydro_params p, int step) {
     }
   }
 
-  double overall_max, overall_min;
+  float overall_max, overall_min;
 
   overall_max = reduce_max(fmax, p);
   overall_min = reduce_min(fmin, p);
@@ -132,7 +132,7 @@ void histo_field(double ***field, hydro_params p, int step) {
 
   int nbins = 100;
 
-  double df = (overall_max - overall_min)/((double)nbins);
+  float df = (overall_max - overall_min)/((float)nbins);
 
   if(fabs(df) < 0.001) {
     printf0(p, "Max: %lf, Min: %lf, df: %lf\n",
@@ -182,29 +182,29 @@ void histo_field(double ***field, hydro_params p, int step) {
     fp = fopen(histodest, "w");
 
     for(i=0; i<nbins; i++) {
-      fprintf(fp, "%lf %d\n", overall_min + df*((double)i), count[i]);
+      fprintf(fp, "%lf %d\n", overall_min + df*((float)i), count[i]);
     }
 
     fclose(fp);
   }
 
-  double end = clock();
+  float end = clock();
 
   printf0(p, "Histogram stuff took %lf\n",
-	  ((double) (end - start)) / CLOCKS_PER_SEC);
+	  ((float) (end - start)) / CLOCKS_PER_SEC);
 
 	       
 }
 
 
-/* void didj(double *cpts, hydro_fields f, hydro_params p)
+/* void didj(float *cpts, hydro_fields f, hydro_params p)
  *
  * *Average* components of the stress-energy tensor for the scalar
  * field to leading order in the metric perturbation.
  */
-void didj(double *cpts, hydro_fields f, hydro_params p) {
+void didj(float *cpts, hydro_fields f, hydro_params p) {
 
-  double cpts_here[TENSOR_CPTS];
+  float cpts_here[TENSOR_CPTS];
 
   int x, y, z;
   int i;
@@ -251,7 +251,7 @@ void didj(double *cpts, hydro_fields f, hydro_params p) {
 
 
   for(i=0; i<TENSOR_CPTS; i++) {
-    cpts[i] = reduce_sum(cpts_here[i], p)/((double)(p.Lx*p.Ly*p.Lz));
+    cpts[i] = reduce_sum(cpts_here[i], p)/((float)(p.Lx*p.Ly*p.Lz));
   }
 
 }
